chore(#647): extract bracketedConsoleLog + add lint guard (A+ hardening)
Move `bracketedConsoleLog` from `phase-executor.ts` to a new
`workflow/notice.ts` module so `run-orchestrator.ts` doesn't import
from a sibling executor (orchestrator → executor was a backwards
dependency direction). Both files now import from `./notice.js`;
test's dynamic import updated to match.

Add a scoped ESLint `no-restricted-syntax` rule on
`src/lib/workflow/phase-executor.ts` that bans `console.log` /
`console.info` / `console.warn` / `console.error` and
`process.stdout/stderr.write`, with the error message pointing at
`bracketedConsoleLog` from `./notice.ts`. Catches future drift —
any new retry message added to this file with a raw `console.log`
fails CI instead of silently re-introducing the scrollback leak.

Two existing `process.stdout/stderr.write` sites in the verbose
streaming callback (lines 676, 687) annotated with
`eslint-disable-next-line` + rationale: `spinner?.pause()` brackets
them above, so they intentionally bypass log-update.

Rule intentionally not applied to `run-orchestrator.ts`: its 10
remaining `console.log` sites are pre-flight (before any issue is
registered with the renderer) or post-dispose, where the live zone
is empty and `console.log` is correct. The two batch-boundary sites
that *do* fire in-window were bracketed in the previous commit.
Forcing the rule there would mean ~10 eslint-disable comments for
zero real catches.

Full suite: 3415 passed, 2 expected fail (negative controls).

Entire-Checkpoint: 1181a179f669
